Started adding constraint size support in the measurement phase of the new UI system. Enabled arithmetic over/underflow detection when compiling the Engine and Simulation projects in debug. Added an OrientationStrategy class and other utility methods to enable algorithm which work with horizontal and vertical directions.

+++ Sources/Engine/Graphics/Texture.cs
@@ -212,7 +212,7 @@ namespace Orion.Engine.Graphics
                 throw new ArgumentException("Invalid pixel region.");
             Argument.EnsureNotNull(pixelData, "data");
 
-            ValidatePixelBufferSize(pixelData, region.Area, pixelFormat);
+            ValidatePixelBufferSize(pixelData, (int)region.Area, pixelFormat);
 
             int lastID;
             GL.GetInteger(GetPName.Texture2D, out lastID);
